+++ assets/src/components/activities/file_upload/schema.ts
@@ -1,0 +1,16 @@
+import { Part, Transformation, ActivityModelSchema, Stem } from '../types';
+
+export interface FileSpec {
+  accept: string;
+  maxCount: number;
+}
+
+export interface FileUploadSchema extends ActivityModelSchema {
+  stem: Stem;
+  fileSpec: FileSpec;
+  authoring: {
+    parts: Part[];
+    transformations: Transformation[];
+    previewText: string;
+  };
+}